requirements.txt
amqp==5.3.1
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.2.1
Flask==3.1.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.2
moviepy==2.2.1
numpy==2.3.0
packaging==25.0
peewee==3.18.1
pillow==11.2.1
proglog==0.1.12
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-mpv==1.0.8
redis==6.2.0
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
yt-dlp==2025.6.9
